requirements.txt
Flask
Flask_SQLAlchemy
Pillow
flask-cors
python-dotenv
gunicorn
Flask-Migrate
flask-jwt-extended
